refactor(transports): split streaming.ts into src/transports/ + core/transport.ts
Split the 1620-line streaming.ts god file into the transport domain:
- core/transport.ts — StreamRequestOptions / TransportRequestSummary types
- transports/chatCompletions.ts / responses.ts / anthropic.ts / google.ts —
  one entry per transport (OpenAI chat / OpenAI Responses / Anthropic
  Messages / Google GenerateContent)
- transports/engine.ts — shared HTTP+SSE streaming engine + retry/backoff
- transports/sse.ts — pure SSE data-line parser
- transports/extractors.ts — Base/OpenAi/Anthropic response extractors
- transports/extract.ts — non-stream extraction + pure delta helpers
- transports/streamParts.ts — progress/thinking part emission
- transports/thinkTags.ts — pure inline <think> tag stripper

streaming.ts is kept as a thin barrel re-exporting the historical public
API. Behavior-preserving; verified with npm run test-retry (mock server).
