+++ src/transports/sse.ts
@@ -1,0 +1,31 @@
+import type * as vscode from "vscode";
+
+/** Pure SSE `data:` line parser — one event string in, extracted parts out. */
+export function parseServerSentEvent(
+  event: string,
+  extractParts: (data: unknown) => vscode.LanguageModelResponsePart[],
+  onData?: (data: unknown) => void,
+): vscode.LanguageModelResponsePart[] {
+  const lines = event
+    .split(/\r?\n/)
+    .filter((line) => line.startsWith("data:"))
+    .map((line) => line.slice("data:".length).trim());
+
+  const parts: vscode.LanguageModelResponsePart[] = [];
+
+  for (const line of lines) {
+    if (!line || line === "[DONE]") {
+      continue;
+    }
+
+    try {
+      const data = JSON.parse(line) as unknown;
+      onData?.(data);
+      parts.push(...extractParts(data));
+    } catch {
+      // Ignore malformed SSE lines; the API may send comments or keep-alive frames.
+    }
+  }
+
+  return parts;
+}
